* master:
  Update the AppConfiguration.Logger obsolete message (#2495)
  Add more serializer types to the preserved ones (#2489)
  Update changelog from Core (#2488)
  Build an xcframework for iOS (#2475)
  Prepare for vNext (#2487)
  Prepare for 10.2.1 (#2484)
  Fix TableKey / ObjKey conversion on Android x86 (#2477)
  Update the build manager locations (#2474)
  Clean up native resources on Unity application exit (#2467)
  Fix some native warnings (#2483)
  Don't try to get the app from a removed user (#2480)
  Use delete_files from core when deleting a Realm. (#2401)
  Fixes for the sync datatype tests (#2461)
  Clean up the package.json and add some docs (#2452)
  Run iOS tests on CI (#2405)
  Verify that objects do not belong to different realm when added to collection (#2465)
  Updated README instructions (#2459)

+++ wrappers/src/results_cs.cpp
@@ -135,7 +135,7 @@ REALM_EXPORT Results* results_get_filtered_results(const Results& results, uint1
 
         std::vector<Mixed> mixed_args;
         mixed_args.reserve(args_count);
-        for (int i = 0; i < args_count; ++i) {
+        for (size_t i = 0; i < args_count; ++i) {
             if (arguments[i].type != realm_value_type::RLM_TYPE_LINK) {
                 mixed_args.push_back(from_capi(arguments[i]));
             }
